+++ Sources/BaseFoundationiOS/Templates/Enums/LoadingState.swift
@@ -1,0 +1,14 @@
+//
+//  LoadingState.swift
+//
+//
+//  Created by pablo borquez on 09/01/24.
+//
+
+import Foundation
+
+public enum LoadingState {
+    case loading
+    case error
+    case loaded
+}