+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # master
 
 - Dictionary `get` funcs added.
+- `printBaseRESTAPI` set to `public` on `OSLogger`. 
 
 # 0.11.0