+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Dictionary `get` funcs added.
 - `printBaseRESTAPI` set to `public` on `OSLogger`. 
 - `TokenType` implemented.
+- `EmailRequired` implemented.
 
 # 0.11.0